test: Extract deployframework methods into helper functions
